nfdiv-1088: move notification sent confirmation to AOS class
